refactor: move sync/pipeline.py to run/pipeline.py
Pipeline (check_and_run, Strictness, PipelineResult) is about
"how to run backups" (compose preflight + sync), not "how to
execute syncs" (rsync + runner). Move to run/ package.

- sync/ keeps rsync.py, runner.py, output.py (sync execution)
- run/ gains pipeline.py (backup orchestration)
- Remove pipeline re-exports from sync/__init__.py
- Update all consumers and test patch strings
