+++ tests/e2e_docker/test_pipeline.py
@@ -38,7 +38,7 @@ from nbkp.config import (
 from nbkp.remote.resolution import resolve_all_endpoints
 from nbkp.remote.testkit.docker import REMOTE_BACKUP_PATH
 from nbkp.orchestration import managed_mount
-from nbkp.sync.pipeline import Strictness, check_and_run
+from nbkp.run.pipeline import Strictness, check_and_run
 from nbkp.sync.runner import SyncOutcome
 
 from nbkp.sync.testkit.seed import build_chain_config
