+++ tests/test_cli.py
@@ -653,8 +653,8 @@ class TestCheckCommand:
 
 
 class TestRunCommand:
-    @patch("nbkp.sync.pipeline.run_all_syncs")
-    @patch("nbkp.sync.pipeline.check_all_syncs")
+    @patch("nbkp.run.pipeline.run_all_syncs")
+    @patch("nbkp.run.pipeline.check_all_syncs")
     @patch("nbkp.cli.common.load_config")
     def test_successful_run(
         self,
@@ -686,8 +686,8 @@ class TestRunCommand:
         assert callable(call_kwargs.kwargs.get("on_sync_start"))
         assert callable(call_kwargs.kwargs.get("on_sync_end"))
 
-    @patch("nbkp.sync.pipeline.run_all_syncs")
-    @patch("nbkp.sync.pipeline.check_all_syncs")
+    @patch("nbkp.run.pipeline.run_all_syncs")
+    @patch("nbkp.run.pipeline.check_all_syncs")
     @patch("nbkp.cli.common.load_config")
     def test_displays_status_before_results(
         self,
@@ -719,8 +719,8 @@ class TestRunCommand:
         ok_pos = result.output.index("OK")
         assert vol_pos < ok_pos
 
-    @patch("nbkp.sync.pipeline.run_all_syncs")
-    @patch("nbkp.sync.pipeline.check_all_syncs")
+    @patch("nbkp.run.pipeline.run_all_syncs")
+    @patch("nbkp.run.pipeline.check_all_syncs")
     @patch("nbkp.cli.common.load_config")
     def test_failed_run(
         self,
@@ -748,8 +748,8 @@ class TestRunCommand:
         assert result.exit_code == 1
         assert "FAILED" in result.output
 
-    @patch("nbkp.sync.pipeline.run_all_syncs")
-    @patch("nbkp.sync.pipeline.check_all_syncs")
+    @patch("nbkp.run.pipeline.run_all_syncs")
+    @patch("nbkp.run.pipeline.check_all_syncs")
     @patch("nbkp.cli.common.load_config")
     def test_dry_run(
         self,
@@ -782,8 +782,8 @@ class TestRunCommand:
         check_kwargs = mock_checks.call_args
         assert check_kwargs.kwargs.get("dry_run") is True
 
-    @patch("nbkp.sync.pipeline.run_all_syncs")
-    @patch("nbkp.sync.pipeline.check_all_syncs")
+    @patch("nbkp.run.pipeline.run_all_syncs")
+    @patch("nbkp.run.pipeline.check_all_syncs")
     @patch("nbkp.cli.common.load_config")
     def test_json_output(
         self,
@@ -825,8 +825,8 @@ class TestRunCommand:
         call_kwargs = mock_run.call_args
         assert call_kwargs.kwargs.get("on_rsync_output") is None
 
-    @patch("nbkp.sync.pipeline.run_all_syncs")
-    @patch("nbkp.sync.pipeline.check_all_syncs")
+    @patch("nbkp.run.pipeline.run_all_syncs")
+    @patch("nbkp.run.pipeline.check_all_syncs")
     @patch("nbkp.cli.common.load_config")
     def test_sync_filter(
         self,
@@ -865,8 +865,8 @@ class TestRunCommand:
         run_kwargs = mock_run.call_args
         assert run_kwargs.kwargs.get("only_syncs") == ["photos-to-nas"]
 
-    @patch("nbkp.sync.pipeline.run_all_syncs")
-    @patch("nbkp.sync.pipeline.check_all_syncs")
+    @patch("nbkp.run.pipeline.run_all_syncs")
+    @patch("nbkp.run.pipeline.check_all_syncs")
     @patch("nbkp.cli.common.load_config")
     def test_progress(
         self,
@@ -903,8 +903,8 @@ class TestRunCommand:
         call_kwargs = mock_run.call_args
         assert call_kwargs.kwargs.get("progress") == ProgressMode.PER_FILE
 
-    @patch("nbkp.sync.pipeline.run_all_syncs")
-    @patch("nbkp.sync.pipeline.check_all_syncs")
+    @patch("nbkp.run.pipeline.run_all_syncs")
+    @patch("nbkp.run.pipeline.check_all_syncs")
     @patch("nbkp.cli.common.load_config")
     def test_exits_before_syncs_on_status_error(
         self,
@@ -922,8 +922,8 @@ class TestRunCommand:
         assert result.exit_code == 1
         mock_run.assert_not_called()
 
-    @patch("nbkp.sync.pipeline.run_all_syncs")
-    @patch("nbkp.sync.pipeline.check_all_syncs")
+    @patch("nbkp.run.pipeline.run_all_syncs")
+    @patch("nbkp.run.pipeline.check_all_syncs")
     @patch("nbkp.cli.common.load_config")
     def test_sentinel_only_proceeds_by_default(
         self,
@@ -950,8 +950,8 @@ class TestRunCommand:
         assert result.exit_code == 0
         mock_run.assert_called_once()
 
-    @patch("nbkp.sync.pipeline.run_all_syncs")
-    @patch("nbkp.sync.pipeline.check_all_syncs")
+    @patch("nbkp.run.pipeline.run_all_syncs")
+    @patch("nbkp.run.pipeline.check_all_syncs")
     @patch("nbkp.cli.common.load_config")
     def test_sentinel_only_exits_when_strict(
         self,
